Interface
Jeg har ændret de forskellige sider så nogle af dem er blevet til frames
fremfor windows, men det blev for besværligt at gøre det når brugens
stemme skulle afgives (jeg gav i hvert fald op på at prøve), så der
åbner der sig bare et nyt vindue :)
