--- AauVotingSystemP4/AauVotingSystemP4/CitizenAboutToVote.xaml
@@ -1,24 +1,0 @@
-﻿<Window x:Class="AauVotingSystemP4.CitizenAboutToVote"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:AauVotingSystemP4"
-        mc:Ignorable="d"
-        Title="CitizenAboutToVote" Height="300" Width="300">
-    <DockPanel Margin="0,0,-0.4,-83.6">
-        <local:HeaderUserControl x:Name="HeaderUserControl" Grid.Row="0" UseLayoutRounding="False" DockPanel.Dock="Top" Height="41.2"/>
-        <StackPanel Orientation="Vertical" Margin="0,0.4,0,74.6">
-            <StackPanel Margin="0,25.4,0,20" Width="295" HorizontalAlignment="Center" Height="40">
-                <TextBlock HorizontalAlignment="Center">
-                "Username" will vote for "selected candidate or party". <LineBreak/>
-                Do you agree?
-                </TextBlock>
-            </StackPanel>
-            <StackPanel Height="38" Orientation="Horizontal" HorizontalAlignment="Center">
-                <Button x:Name="No" Content="No" Height="27" Width="94" Click="No_Click" Margin="0,0,20,0.8"/>
-                <Button x:Name="Yes" Content="Yes" Height="26" Width="87" Click="Yes_Click"  Margin="20,0,0,0"/>
-            </StackPanel>
-        </StackPanel>
-    </DockPanel>
-</Window>
