+++ AauVotingSystemP4/AauVotingSystemP4/CitizenHaveVoted.xaml
@@ -1,0 +1,23 @@
+﻿<Window x:Class="AauVotingSystemP4.CitizenHaveVoted"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:AauVotingSystemP4"
+        mc:Ignorable="d"
+        Title="CitizenHaveVoted" Height="300" Width="450">
+    <DockPanel>
+        <local:HeaderUserControl x:Name="HeaderUserControl" Grid.Row="0" UseLayoutRounding="False" DockPanel.Dock="Top" Height="40"/>
+        <StackPanel Orientation="Vertical" Margin="0,0.4,0,74.6" Width="309">
+            <StackPanel Margin="-49,25.4,-49.2,20" Width="407" HorizontalAlignment="Center" Height="40">
+                <TextBlock HorizontalAlignment="Center" Margin="51,0,10.2,0" Width="346">
+                "Username" has voted for "selected candidate or party". <LineBreak/>
+                You can return to the homepage by clicking the button bellow.
+                </TextBlock>
+            </StackPanel>
+            <StackPanel Height="35" Orientation="Horizontal" HorizontalAlignment="Center">
+                <Button x:Name="Homepage" Content="Homepage" Height="27" Width="94" Click="Homepage_Click" />
+            </StackPanel>
+        </StackPanel>
+    </DockPanel>
+</Window>
