+++ AauVotingSystemP4/AauVotingSystemP4/Citizen_Homepage.xaml
@@ -1,0 +1,17 @@
+﻿<Page x:Class="AauVotingSystemP4.Citizen_Homepage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:AauVotingSystemP4"
+      mc:Ignorable="d" 
+      d:DesignHeight="300"
+      Title="Citizen_Homepage" Width="486.538">
+
+    <Grid>
+        <Button x:Name="SeePreviouslyElections" Content="See previously elections" HorizontalAlignment="Left" Margin="280,50,0,0" VerticalAlignment="Top" Width="117" Click="SeePreviouslyElections_Click" />
+        <Button x:Name="VoteButton" Content="Vote" HorizontalAlignment="Left" Margin="179,50,0,0" VerticalAlignment="Top" Width="75" RenderTransformOrigin="0.449,1.222" Click="VoteButton_Click"/>
+        <Button x:Name="SnydeKnap" Content="Psst tryk på mig og kom til ElectionBoard" HorizontalAlignment="Left" Margin="10,10,0,0" VerticalAlignment="Top" Width="231" Click="SnydeKnap_Click"/>
+        <Frame Name="Main" Margin="0,100,0,0"/>
+    </Grid>
+</Page>
