--- AauVotingSystemP4/AauVotingSystemP4/ElectionboardOptions.xaml
@@ -1,13 +1,0 @@
-﻿<UserControl x:Class="AauVotingSystemP4.ElectionboardOptions"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:AauVotingSystemP4"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" Width="577.814">
-    <Grid>
-        <Button x:Name="ParliamentElection" Content="Parliament election" HorizontalAlignment="Left" Margin="280,50,0,0" VerticalAlignment="Top" Width="117" Click="ParliamentElection_Click"/>
-        <Button x:Name="EUElection" Content="EUElection" HorizontalAlignment="Left" Margin="179,50,0,0" VerticalAlignment="Top" Width="75" RenderTransformOrigin="0.449,1.222"/>
-    </Grid>
-</UserControl>
