+++ AauVotingSystemP4/AauVotingSystemP4/ElectionboardVotingBallot.xaml.cs
@@ -33,7 +33,8 @@ namespace AauVotingSystemP4
 
         private void Delete_Click(object sender, RoutedEventArgs e)
         {
-
+            DeleteVotingOption dvo = new DeleteVotingOption();
+            dvo.Show();
         }
     }
 }
