--- AauVotingSystemP4/AauVotingSystemP4/SeeElections.xaml.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace AauVotingSystemP4
-{
-    /// <summary>
-    /// Interaction logic for SeeElections.xaml
-    /// </summary>
-    public partial class SeeElections : Window
-    {
-        public SeeElections()
-        {
-            InitializeComponent();
-        }
-    }
-}
