+++ AauVotingSystemP4/AauVotingSystemP4/SeePreviouslyElections.xaml
@@ -1,10 +1,13 @@
-﻿<UserControl x:Class="AauVotingSystemP4.PreviouslyElections"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:AauVotingSystemP4"
-             mc:Ignorable="d" d:DesignWidth="300" Height="295.294">
+﻿<Page x:Class="AauVotingSystemP4.SeePreviouslyElections"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:AauVotingSystemP4"
+      mc:Ignorable="d" 
+      d:DesignHeight="300"
+      Title="SeePreviouslyElections" Width="256.731">
+
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
@@ -22,4 +25,4 @@
         </ListBox>
         <Button Grid.Column="0" Grid.Row="2" Margin="0,4.8,0,118.2" Width="125" HorizontalAlignment="Right" Click="Button_Click">View</Button>
     </Grid>
-</UserControl>
+</Page>
